requirements.txt
schedule==1.2.2
oauth2client==4.1.3
oauthlib==3.2.2
google-auth==2.30.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
gspread-formatting==1.1.2
google-api-python-client==2.126.0
python-telegram-bot==13.15
